COG-1197: NoUpdateExecution is moved to OptionalExecution
